Remove obsolete MsSql2008 dialect
