+++ NHibernate.Spatial.sln
@@ -21,8 +21,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests.NHibernate.Spatial.Po
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NHibernate.Spatial.MsSql", "NHibernate.Spatial.MsSql\NHibernate.Spatial.MsSql.csproj", "{3BA49B38-653B-4B24-8F2D-3416B07AAA5D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests.NHibernate.Spatial.MsSql2008", "Tests.NHibernate.Spatial.MsSql2008\Tests.NHibernate.Spatial.MsSql2008.csproj", "{D32A19AC-F7EE-4CD1-AA01-BE8B171B37A0}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests.NHibernate.Spatial.MsSql2012", "Tests.NHibernate.Spatial.MsSql2012\Tests.NHibernate.Spatial.MsSql2012.csproj", "{12E29E47-4760-427E-8EF9-5D2F39B0E980}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8DBF64ED-50C5-42C0-AED7-64FD617FB649}"
@@ -72,10 +70,6 @@ Global
 		{3BA49B38-653B-4B24-8F2D-3416B07AAA5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3BA49B38-653B-4B24-8F2D-3416B07AAA5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3BA49B38-653B-4B24-8F2D-3416B07AAA5D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D32A19AC-F7EE-4CD1-AA01-BE8B171B37A0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D32A19AC-F7EE-4CD1-AA01-BE8B171B37A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D32A19AC-F7EE-4CD1-AA01-BE8B171B37A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D32A19AC-F7EE-4CD1-AA01-BE8B171B37A0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{12E29E47-4760-427E-8EF9-5D2F39B0E980}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{12E29E47-4760-427E-8EF9-5D2F39B0E980}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{12E29E47-4760-427E-8EF9-5D2F39B0E980}.Release|Any CPU.ActiveCfg = Release|Any CPU
