--- Tests.NHibernate.Spatial.MsSql2008/MsSql2008CriteriaFixture.cs
@@ -1,14 +1,0 @@
-﻿using NHibernate.Cfg;
-using NUnit.Framework;
-
-namespace Tests.NHibernate.Spatial
-{
-    [TestFixture]
-    public class MsSql2008CriteriaFixture : CriteriaFixture
-    {
-        protected override void Configure(Configuration configuration)
-        {
-            TestConfiguration.Configure(configuration);
-        }
-    }
-}
