--- Tests.NHibernate.Spatial.MsSql2008/appsettings.json
@@ -1,7 +1,0 @@
-{
-  "ConnectionStrings": {
-    "_comment1": "NOTE: Must specify SQL Server 2012 rather than SQL Server 2008 for the Type System Version attribute (see issue #72)",
-    "_comment2": "NOTE: TrustServerCertificate=True required after updating from System.Data.Sql to Microsoft.Data.Sql as Encrypt is now set to true by default; see: https://stackoverflow.com/a/17658821/3628232",
-    "MsSql2008": "Server=(local)\\MSSQLSERVER2008;Initial Catalog=nhsp_test;User ID=nhsp_test;Password=nhsp_test;Type System Version=SQL Server 2012;TrustServerCertificate=True"
-  }
-}
