+++ Tests.NHibernate.Spatial.MsSql2012/NtsTestCases/Data/vivid/TestSimple.xml
