+++ Tests.NHibernate.Spatial.MsSql2012/appsettings.json
@@ -1,6 +1,7 @@
 {
   "ConnectionStrings": {
-    "_comment": "NOTE: TrustServerCertificate=True required after updating from System.Data.Sql to Microsoft.Data.Sql as Encrypt is now set to true by default; see: https://stackoverflow.com/a/17658821/3628232",
+    "_comment1": "NOTE: Must specify SQL Server 2012 for the Type System Version attribute (see issue #72)",
+    "_comment2": "NOTE: TrustServerCertificate=True required after updating from System.Data.Sql to Microsoft.Data.Sql as Encrypt is now set to true by default; see: https://stackoverflow.com/a/17658821/3628232",
     "MsSql2012": "Server=(local)\\MSSQLSERVER2012;Initial Catalog=nhsp_test;User ID=nhsp_test;Password=nhsp_test;Type System Version=SQL Server 2012;TrustServerCertificate=True"
   }
 }
